refactor FileTransfer and it is a method of service side and client side
