+++ share/share.go
@@ -23,6 +23,9 @@ const (
 	OpencensusSpanClientKey = "opencensus_span_client_key"
 	// OpencensusSpanRequestKey span key in request meta
 	OpencensusSpanRequestKey = "opencensus_span_request_key"
+
+	// SendFileServiceName file transfer service.
+	SendFileServiceName = "_filetransfer"
 )
 
 var (
@@ -49,3 +52,21 @@ var ReqMetaDataKey = ContextKey("__req_metadata")
 
 // ResMetaDataKey is used to set metatdata in context of responses.
 var ResMetaDataKey = ContextKey("__res_metadata")
+
+// FileTransferArgs args from clients.
+type FileTransferArgs struct {
+	FileName string            `json:"file_name,omitempty"`
+	FileSize int64             `json:"file_size,omitempty"`
+	Meta     map[string]string `json:"meta,omitempty"`
+}
+
+// FileTransferReply response to token and addr to clients.
+type FileTransferReply struct {
+	Token []byte `json:"token,omitempty"`
+	Addr  string `json:"addr,omitempty"`
+}
+
+// DownloadFileArgs args from clients.
+type DownloadFileArgs struct {
+	FileName string `json:"file_name,omitempty"`
+}
